// src/schema/upload.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Upload: any;
};

export type File = {
  __typename?: "File";
  name: Scalars["String"];
  src: Scalars["String"];
};

export type Files = {
  __typename?: "Files";
  files: Array<Maybe<File>>;
};

export type FilesInput = {
  files: Array<Scalars["Upload"]>;
  names: Array<Scalars["String"]>;
};